Element by id rather than by object (hash) id
During the work of the websocket API and the webeditor, a strange bug surfaced.
The random path generator throwed an exception, saying it could not find any
valid path from an element. It turned out that the element was in the
Model.elementByElementCache, but with a different hash id than expected.
It seams that the Model.elementByElementCache was not updated when needed.

So, rather that depending on elements hash id (which needs to be generated, and
updated from time to time), I changed the elementByElementCache to
elementByIdCache. Since the id must be unique, this should be good enough.